Varnish: Fix #14 Remove health check, prevent 503
When something is wrong with the backend the Varnish health check (will call pub/health_check.php) will block of any request until that health check is now green again.

This prevents stampeding of the backend and that is good in production, but in development we just want an error.