Remove NFS volumes for MySQL and ElasticSearch
This will make it use guest-only volumes for optimal performance. This
means that MySQL data can not be included in backups.

This will not affect any existing projects, until the changes are
applied from this example file. This should be done on a per-project
basis, and requires developers of each proejct to manually remove
the old volumes before attempting to start docker-composer with the new
setup.